Don't clear adoptedStyleSheets when moving shadow root between document <-> template
As discussed in https://github.com/WICG/construct-stylesheets/issues/133,
it should be safe to keep the adoptedStyleSheets when the shadow root
is moved between a document and a <template> within that document. We
should still clear the adoptedStyleSheets when moving a shadow root
between documents outside of that case (e.g. between iframes).

Bug: 1111864
Change-Id: I803e75e6339f9228289188e4ed7d3f14c6e2432c