[selectors4] Implement :dir pseudo-class
This patch adds support for the new ":dir" pseudo-class.

The patch is covered by several of tests imported from the web tests
including generic tests in addition to the tests from Firefox and
apart from some minor changes on current tests to add the new selector.

Intent-to-ship thread is available at:
https://groups.google.com/u/1/a/chromium.org/g/blink-dev/c/p0Wc66rbVOc

BUG=576815

Change-Id: I503c6af95d9c5817e3e71b1b2025a5562929f0d8